requirements.txt
Django==4.2.*
django-statici18n==2.3.*
django_webpack_loader==1.7.*
django_picklefield==3.1
django_formtools==2.4
django_simple_captcha==0.6.*
djangosaml2==1.5.*
djangorestframework==3.14.*
python-dateutil==2.8.*
pyjwt==2.6.*
pycryptodome==3.20.*
python-cas==1.6.*
pysaml2==7.2.*
requests==2.31.*
requests_oauthlib==1.3.*
future==0.18.*
gunicorn==20.1.*
mysqlclient==2.1.*
qrcode==7.3.*
pillow==10.2.*
chardet==5.1.*
cffi==1.15.1
captcha==0.5.*
openpyxl==3.0.*
Markdown==3.4.*
bleach==5.0.*
python-ldap==3.4.*
pypinyin==0.50.*
pyheif-pillow-opener==0.1.0
